Simplify code
Linear search is techincally slower than the previous binary search,
but it hardly matters for such a short array. The new code is much
easier to understand and debug. It also reads the bit length from
the table instead of hardcoding it, which seems to be the original
intention of the file format.